requirements.txt
Django==4.2.3
djangorestframework==3.14.0
psycopg2>=2.8.6,<2.9
django-mysql==3.9
django-cors-headers==3.5.0
pika==1.1.0
python-dotenv==0.21.0